Shoot

+++ VS/ConsolenSpiel/ConsolenSpiel/TextBuffer.h
@@ -4,13 +4,13 @@
 class TextBuffer
 {
 public:
-	unsigned Width, Height;
+	short Width, Height;
 
 	char * field = nullptr;
-	TextBuffer(unsigned width, unsigned height, char * map);
+	TextBuffer(short width, short height, char * map);
 	~TextBuffer();
-	void SetChar(unsigned x, unsigned y, char character);
-	char GetChar(unsigned x, unsigned y);
+	void SetChar(short x, short y, char character);
+	char GetChar(short x, short y);
 	void Render();
 };
